requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
db-dtypes==1.2.0
discord==2.3.2
discord.py==2.3.2
disnake==2.9.1
exceptiongroup==1.2.0
Flask==3.0.3
Flask-Cors==4.0.0
Flask-DebugToolbar==0.14.1
fpdf==1.7.2
frozenlist==1.4.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.20.1
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
gspread==6.1.0
h11==0.14.0
httplib2==0.22.0
idna==3.5
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
MouseInfo==0.1.3
multidict==6.0.5
numpy==1.24.4
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.0.3
pandas-gbq==0.22.0
Pillow==10.1.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.21
pydata-google-auth==1.8.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
python3-xlib==0.15
pytweening==1.0.7
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
StrEnum==0.4.15
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.2
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
